[IDLE-61] refactor#82: AuthenticationService 검증 로직 분리
	- 코드 리뷰 반영 (p3: 각 예외처리를 if절을 사용하여 처리하는 것보다는 각각 함수로 빼서 처리)
	- SignUpResponse를 presentation단의 dto로 이동
