+++ user-service/src/main/java/kr/mybrary/userservice/authentication/domain/AuthenticationService.java
@@ -1,6 +1,6 @@
 package kr.mybrary.userservice.authentication.domain;
 
-import kr.mybrary.userservice.authentication.domain.dto.SignUpResponse;
+import kr.mybrary.userservice.authentication.presentation.dto.response.SignUpResponse;
 import kr.mybrary.userservice.authentication.presentation.dto.request.SignUpRequest;
 
 public interface AuthenticationService {
