+++ user-service/src/main/java/kr/mybrary/userservice/authentication/presentation/dto/response/SignUpResponse.java
@@ -1,4 +1,4 @@
-package kr.mybrary.userservice.authentication.domain.dto;
+package kr.mybrary.userservice.authentication.presentation.dto.response;
 
 import kr.mybrary.userservice.user.persistence.Role;
 import lombok.Builder;
